server: Rename opts fields/add known GCE CertPools
