+++ server/policy_constants.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"strconv"
 
+	"github.com/google/certificate-transparency-go/x509"
 	pb "github.com/google/go-tpm-tools/proto/attest"
 )
 
@@ -63,10 +64,39 @@ var (
 // Known Privacy CA certs.
 var (
 	//go:embed ca-certs/tpm_ek_root_1.cer
-	GceEKRootCA []byte
+	gceEKRootCA []byte
 	//go:embed ca-certs/tpm_ek_intermediate_2.crt
-	GceEKIntermediateCA2 []byte
+	gceEKIntermediateCA2 []byte
 )
+
+// CertPools corresponding to the known CA certs.
+var (
+	GceEKRoots         *x509.CertPool
+	GceEKIntermediates *x509.CertPool
+)
+
+func init() {
+	var err error
+	GceEKRoots, err = getPool([][]byte{gceEKRootCA})
+	if err != nil {
+		panic("failed to create the root cert pool, because some certs failed to parse")
+	}
+	GceEKIntermediates, err = getPool([][]byte{gceEKIntermediateCA2})
+	if err != nil {
+		panic("failed to create the intermediate cert pool, because some certs failed to parse")
+	}
+}
+func getPool(certs [][]byte) (*x509.CertPool, error) {
+	pool := x509.NewCertPool()
+	for _, certBytes := range certs {
+		cert, err := x509.ParseCertificate(certBytes)
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse cert: %v", err)
+		}
+		pool.AddCert(cert)
+	}
+	return pool, nil
+}
 
 // ConvertSCRTMVersionToGCEFirmwareVersion attempts to parse the Firmware
 // Version of a GCE VM from the bytes of the version string of the SCRTM. This
